building: implement support for building with macOS universal2 bootloaders
Before appending data package to macOS bootloader, we need to ensure
that it is not signed. Typically, this is true for x86_64 bootloaders
(thin binaries or slices in universal2 fat binaries). However, arm64
bootloaders (thin binaries or slices in universal2 fat binaries) do
contain an ad-hoc signature due to stricter signature requirements on
Apple M1 Silicon platform. Therefore, if present, the signature needs
to be removed from the last slice (which, in universal2 binary, is arm64)
and then re-generated once the package data is appended.

However, it seems that when cross-compiling universal2 bootloader with
osxcross, we end up with both x86_64 and arm64 slice signed. While
the signature on the former should not have any impact, it does seem
to cause issues when trying to sign the final binary using the
real identity (codesign returns "invalid or unsupported format for
signature"). Therefore, we strip signatures from all slices, and
after data is appended, re-generate ad-hoc signature on all slices.
These ad-hoc signatures on the final binary can serve as integrity
validation in cases when users do not provide their own signature.

Add a new option to EXE(), called target_arch, which currently
applies only to macOS (but could have wider uses, hence generic
name). Specifying target architecture (x86_64, arm64, universal2)
allows building single-arch thin-binary frozen applications from
universal2 python and packages/libraries. It is also used to
check for invalid combinations (requesting universal2 build with
single-arch bootloader, or requesting architecture that is
unavailable in the bootloader).

By default, the target architecture corresponds to the current
running architecture (i.e., universal2 build needs to be explicitly
enabled due to potential issues).

For now, the option needs to be manually specified as an argument
to EXE() in the spec file. At the moment, it also applies only
to the final executable but not to collected binaries.