cliutils: archive_viewer: fix crash when quitting or moving up a level
Calling arch.lib.close() when cleaning up the archive when either
moving up a level or quitting the archive_viewer application
triggers the assert in pyimod02_archive.ArchiveFile.__getattr__.

This is because local file object is set only between __enter__
and __exit__ calls, i.e., while in the `with arch.lib:` block.
Which also means that there's no need for attempting to close
those file handles in the first place, and the error is probably
a regression from when the thread-local file objects were
introduced.