bootloader: clean up endianess conversion (#5457)
* bootloader: perform endianess fix-up of archive fields in only one place

The conversion of multi-byte integer fields in COOKIE and TOC
structures is a data (de)serialization detail, and as such should
be performed only once, after the data is read; as opposed to
on-the-fly endianess conversions on every field access, which are
scattered throughout the codebase.

* bootloader: on Windows, avoid using ntohl() from ws2_32 library

On Windows, use built-in byte-swap macros instead of ntohl(), which
is available from winsock2 library. Using the latter requires the
bootloader to be linked against the winsock2 library, which may have
other unintended consequences.

On all other platforms, continue using ntohl() for now, but hide it
behind the pyi_be32toh() macro.