archive: fix stored paths when building under MSYS
When building CArchive package under MSYS, os.path.sep is Unix-style
forward clash separator that is incompatible with bootloader (which
on Windows expects back slash separators). Therefore, if we are on
Windows and os.path.sep is forward slash, os.path.normpath() in
CTOC.add() needs to be followed by explicit replacement of forward
slashes with back slashes.

Otherwise, bootloader fails to extract nested files from archive in
onefile builds that are created in MSYS environment.