tests: extend the multipackage tests with nested resources
Extend the multipackage test to verify references to nested resources
(data files and extensions) within the shared package in the
multipackage mode. For nested extension, the psutil package is used
(as it contains one).

Refactor the test scripts to use a common test function from a
package to avoid duplicating test code.

The extended tests shows that multipackage currently does not
properly handle resources that are collected in subdirectories of
_MEIPASS instead of _MEIPASS itself.