bootloader: use Py_UnbufferedStdioFlag to enable unbuffered output (#5597)
* tests: add a test for unbuffered output mode

The frozen script periodically prints * character to a single
line on the specified output stream (stdout or stderr) in either
text or binary mode. Once the selected number of characters have
been printed, end-of-transmission is signalled by 'E' character.

The test itself is asynchronously reading the output of the spawned
frozen script, and counting the recived * characters as long as
the received buffer ends with *. In buffered mode, the whole
E-terminated line is exepected to be received at once, while in
unbuffered mode, each * character is expected to be received
separately.

* bootloader: use Py_UnbufferedStdioFlag to enable unbuffered output

Setting Py_UnbufferedStdioFlag enables unbuffered stdio mode in
the Python library. This enables unbuffered binary layer of
stdout and stderr streams in all supported Python versions, while
unbuffered text layer requires Python 3.7 or later.