building: ensure that onefile builds on Windows have manifest embedded
The onefile builds on Windows may not have manifest embedded if
icon was explicitly disabled via `--icon NONE`. This in turn causes
subtle issues, such as #5624.

Remove the convoluted check for entering the resource-modification
codepath, as we always need to reach the part that embeds the
manifest in the onefile executable, and the rest of
resource-modification parts are already guarded by their corresponding
checks. This should simplify the code and make it less prone to
errors.