building: macOS: prevent checkCache from re-processing cached results
On macOS, we've been re-writing library paths on cached results
as well. This is probably (hopefully?) unnecessary, as those path
adjustments depend only on dist_nm. As path rewriting modifies
headers, this forces us to re-sign the binaries on practicaly every
run, which negates the point of the cache.

However, the strip + upx combination should probably properly
forward dist_nm to its call to checkCache (even if the paths get
subsequently rewritten again).