depend: allow compiled modules to be collected into base_library.zip
This allows PyInstaller to work with embedded python on Windows,
provided that the user worked around the lack of support for
collecting files from python3X.zip (i.e., by replacing the
python3X.zip file with same-named folder that contains unpacked
contents of the zip).

In python 3.7 and later, the collected modules end up with hash
that is derived from the raw contents of the original compiled
module (as opposed to source, which is unavailable). While this
is technically incorrect, it should also be irrelevant in a
frozen application, and is therefore not explicitly addressed
(e.g., by copying hash from the original compiled module).