setup: Build per-platform wheels.
Build separate wheels for each platform containing only
the correct bootloaders for said platform with no C
source code/build scripts. Set the platform tag so that
the correct wheel will be selected automatically by pip.

To facilitate cross compiling (or rather cross platform wheel
building), add new cross-build wheel commands to the setup.py and
a "build the lot" command (bdist_wheels) to run them all:

    $ python setup.py --help-commands
    ...
    Extra commands:
        ...
        wheel_windows_64bit  Create a Windows-64bit wheel
        wheel_windows_32bit  Create a Windows-32bit wheel
        wheel_linux_64bit    Create a Linux-64bit wheel
        wheel_linux_32bit    Create a Linux-32bit wheel
        wheel_darwin_64bit   Create a Darwin-64bit wheel
        ...
        bdist_wheels         Build all available wheel types
        ...

With the above in place, it no longer makes sense to include
bootloaders in source distributions (sdists) as the only time
pip will choose the sdist over the wheels is if it is
installing on a platform which is incompatible with all the
wheels' platforms and therefore all of the pre-built bootloaders.
Therefore, strip all bootloaders (and various other things that
shouldn't have been there anyway) from the MANIFEST.in.