hooks: PySide2/PyQt5: unify handling of relative Qt directory
Extend Qt5LibraryInfo with new property, qt_rel_dir, which denotes
the relative Qt data path in the frozen application. Its value
varies between PyQt5 and PySide2, their versions, and platforms:
* PyQt5/Qt for PyQt5 <= 5.15.3 on all platforms
* PyQt5/Qt5 for PyQt5 >= 5.15.4 on all platforms
* PySide2/Qt for PySide2 on Linux and macOS
* PySide for PySide2 on Windows

This way, we now have a unified, central location when this path
is determined, instead of the corresponding functionality being
scattered across the hooks.

Adds support for PyQt5 5.15.4, and fixes #5631.

The paths used by rthooks also need to be adjusted to handle PyQt5/Qt
vs PyQt5/Qt5 switch, and to handle the difference between PySide2
on Windows and PySide2/Qt on other platforms (as we now make
distinction between the two).