building: splash: improve error messages when Pillow is missing
If Pillow is missing and image needs to be resized, raise a
ValueError instead of None, to avoid a more confusing error
further down the road.

Similarly, if lack of Pillow precludes image conversion, mention
that in the error message to make it clear what needs to be done
to resolve the error.