Depend: Replace old ctypes scanning code with regex-based implementation.
Doing so not only brings it in line with the `pkg_resources`/
`importlib.metadata` bytecode scanning code for determining metadata
dependencies, but it also fixes an unoticed bug where the ctypes scanner
fails if the preciding file contents contain more than 256 literals or
names. In this scenario, Python's bytecode switches to an EXTENDED_ARG
mode which the original implementation was unaware of.