building: adjust macOS SDK version reported by frozen application
Ensure that the macOS SDK version reported by the frozen application
(LC_VERSION_MIN_MACOSX load command in executable's headers)
corresponds to the minimum of the SDK version used to build the
bootloader and the SDK version used to build the Python library
(and by extension, other python-bundled libraries, such as Tcl/Tk).

Having the application report more recent version that Python
library may result in macOS attempting to enable additional features
that are not available in the Python library, which may in turn
case inconsistent behavior and UI issues with `tkinter`. Overall,
matching the SDK version of Python library should ensure the most
consistent behavior w.r.t. python interpreter executable, which
reports that same version.

An example of SDK-version-based feature activation is macOS dark
mode, which was introduced in macOS 10.14 Mojave. Python.org
macOS Intel installers are still built against older 10.9 SDK,
so their Tcl/Tk 8.6.8 libraries do not enable support for dark
mode (and may be further lack support for it, anyway). This is
not a problem when used with python interpreter, which also
reports 10.9 SDK and is in turn opted out from dark mode by the
OS. However, when those Tcl/Tk libraries are used with PyInstaller
bootloader that was built against macOS SDK 10.14 or newer, the
OS activates dark mode, which results in black Tk windows (#5827).

A similar UI issue is present in Anaconda python, which is built
against 10.10 SDK. While their python 3.9 uses Tcl/Tk 8.6.10,
the older SDK again precludes support for dark mode, and using
those Tcl/Tk libraries with PyInstaller bootloader that was
built against macOS SDK 10.14 or newer, the UI ends up with
white fonts on bright background.

In both cases, adjusting the reported SDK version back to 10.9
or 10.10, repsectively, disables the dark mode, and fixes the
UI issues.