hookutils/qt: introduce get_qt_library_info(namespace) helper
Returns the Qt5LibraryInfo that corresponds to the given namespace
(e.g., pyqt5_library_info for 'PyQt5' and pyside2_library_info
for 'PySide2'.

This is primarily intended to allows us write more generic code
in our utility functions, where previously we kept switching on
the namespaces to select the correct *_library_info symbol.

For hooks it still makes more sense to directly import the
*_library_info symbol that they require.