tests: include PySide6 and PyQt6 in applicable Qt tests
Add PySide6 and PyQt6 to QtPyLibs so that they are included in
applicable Qt tests. Rename generic Qt tests to follow the
test_Qt_{area} naming convention. Currently, the generic tests
are:
* test_Qt_QtQml
* test_Qt_QtNetwork_SSL_support
* test_Qt_QTranslate

The rest of the Qt-related tests are either package-specific
(test_PyQt5_Qt, test_PyQt5_uic) and would require some further
generalization, or are not applicable to Qt6 yet (i.e., QtWebEngine
test, due to the module not having been ported to Qt6 yet).