tests: add test_Qt_QtWidgets as basic widget test
This basic widget test serves as a sanity check for the more
complex test_qt_QtQml (so we can determine whether the failure
in the latter is tied to Qml, or caused by a more basic problem,
in which case both tests should fail).

As QtWidgets depends on QtGui and the test code does not explicitly
import the latter, this test also checks whether our dynamic
dependency resolution between Qt modules works correctly.