hooks: add hooks for PyQt6
Based on intersection of existing PyQt5 hooks and modules available
in Qt6. So we don't include hooks for modules that have not been
ported to Qt6 (yet), such as QtMultimedia, QtScript, QtScript, or
QtWebEngine. On the other hand, we do not fully cover all modules
available in Qt6 (e.g., QtOpenGLWidgets, QtSvgWidgets), but neither
is that the case in the original Qt5 hooks.

The implmented set of hooks is enough to make the tests pass, and
should suffice for widget-based and QML-based applications.