hooks: Qt rthooks: add sys._MEIPASS to PATH on Windows
Modify rthooks for PyQt5, PyQt6, PySide2, and PySide6 so that
sys._MEIPASS is added to PATH on Windows.

PyQt5 already had this PATH modification, but it was done on all
OSes, causing issues on macOS when trying to spawn python
interpeter from the frozen application:
https://github.com/pyinstaller/pyinstaller/issues/4859#issuecomment-811046377
Therefore, we now have the PyQt5 rthook now modify PATH only on
Windows.

For PyQt6, the PATH modification seems required for collected SSL
DLLs to be discoevered by QtNewtork in onefile application. Same
goes for PySide2 and PySide6, although there the SSL DLLs need to be
first manually installed in the package directory in order to be
available during collection/application freezing.