tests: qt: work around the problem with imports under pytest.monkeypatch
The PyQt6 and PySide6 tests are failing on Windows because they
perform their first import of the package within the test (as
part of pytest.importorskip() call) with pytest.monkeypatch
being active. When Qt-based package is imported, it adds path to
its DLLs to PATH, so if the import happens under pytest.monkeypatch,
the change is lost for subsequent tests. This in turn causes
incomplete builds of test programs, because the freezing of test
program does not spawn separate pyinstaller process, but rather
calls pyi_main.run() from the test process.

PySide2 and PyQt5 are not affected by this, because we have tests
with @importorskip('PySide2') and @importorskip('PyQt5') annotations,
which cause these packages to be imported when pytest is collecting
tests. Hence their import happens without monkeypatch being active,
and their additions to PATH are permanent.

So to work around this, we manually attempt to import all Qt-based
packages (PySide2, PyQt5, PySide6, and PyQt6), same as if there
was an @importorskip() annotation involving each of them.