hooks: PIL: exclude all Qt-based packages
Extend the excludes list with all Qt-based packages: PySide2,
PySide6, and PyQt6 in addition to already-excluded PyQt5.

Otherwise, test_pil_PyQt5 (or anything using PIL.ImageQt) ends
up collecting PyQt5, as well as  PySide2, PySide6 and PyQt6
(assuming all four are installed).

And because Qt shared libraries on macOS do not contain version
in the name (there's one in framework bundle, which we do not
preserve), we end up with one copy of e.g. QtGui in sys._MEIPASS.
And if that happens to be incompatible with PyQt5 (i.e., from
PySide6 or PyQt6), it leads to segfault, and in turn test failure.