bootloader: macOS: always send messages to syslog if in windowed mode
Instead of sending messages from pyi_global_printf() and
pyi_global_perror() to macOS syslog if bootloader is build in
both windowed and debug mode (i.e., only in runw_d variant),
do it for both windowed variants.

This allows us to log unhandled exception messages to the syslog
regardless of whether the debug mode is enabled or not, and should
give better indication why the app is "silently crashing". The
non-error messages are still disabled for non-debug builds,
as the corresponding macros (e.g., VS()) are turned nop-op
in non-debug bootloaders.