bootloader: improve unhandled exception reporting
Improve the message about failed script execution, by making
it clear that the failure was due to unhandled exception
in Python code. In console mode, this involves rewording
the message.

On Windows in windowed mode, use the new exception dialog to
display information about exception (traceback). On macOS in
windowed mode (app bundles), use FATALERROR() to report the
error and dump the traceback; this sends them to syslog, and
makes them visible to the macOS Console.app.