loader: ctypes: switch the search order for frozen libraries
Modify the _frozen_name() so that it first checks if the given
(file)name exists, before falling back to searching sys._MEIPASS
for it.

This effectively switches the search order: whereas formerly we
would first check sys._MEIPASS and use the given name as fallback,
we now first check the given name and then search sys._MEIPASS as
a fallback.