simplify LD_LIBRARY_PATH_ORIG processing, fixes #2892
if _ORIG is always there, users can easily restore LDLP to its original
value before forking a process. it must be always there so users can
easily differentiate whether their code is running from a pyinstaller
bundle (no matter whether it is single-file or single-directory) or not.

e.g. when forking to the system ssh, it is important that ssh loads the
system openssl lib (not the openssl lib inside the pyinstaller bundle).