Add support for pkgutil.iter_modules (#5959)
* tests: add tests for pkgutil.iter_modules

Test that pkgutil.iter_modules() lists same contents of a package in
unfrozen and frozen version. For tests, we use altgraph (pure python)
and psutil (contains binary extensions), as well as psutil.tests
(as an example of a sub-package).

We also test both archive and noarchive mode, because in the latter,
.pyc modules end up directly on filesystem, and are subject of python's
built-in FileFinder.

There, noarchive variants of tests pass without additional changes,
while archive variants fail as frozen versions do not list PYZ-embedded
modules.

* hooks: add rthook for pkgutil to override iter_modules

Provide custom implementation of iter_modules() that works with
our FrozenImporter and allows listing of sub-modules that are
embedded in the PYZ archive.