Depend: Bytecode scanner: Skip false positive matches starting on odd bytes.
All bytecode instructions come in OPCODE, ARGUMENT pairs of bytes so that
every even byte is an OPCODE and every odd one is an ARGUMENT.

Formally, this wasn't enforced by our bytecode scanner, leading to false
positive matches when a sequence of pairs' ARGUMENTS just happens to be the
sequence of bytes as the OPCODEs which the scanner is looking for. These false
positive matches are likely to become invalid after further parsing which led to
index errors (such as in #6007).

Approximately speaking, the fix is simply to skip any `re.Match` found that does
not satisfy `match.start() % 2 != 0`.

To streamline any other issues regarding bytecode scanning, the top level code
which invokes the bytecode scanner now catches and reraises errors, adding
exactly which module it failed to scan to the error message.