depend: suppress warnings about missing (lib)shiboken*
These are harmless, and are result of the shiboken2/6 shared
library residing in a separate package and not being in the
search path. Therefore, PySide2/6 extensions themselves cannot
resolve the shiboken shared library, which actually ends up
loaded when the corresponding extension from shiboken package is
loaded (and this is also how PyInstaller manages to pick it up).