Tests: Rewrite @importorskip() to not import anything.
The previous version of `@importorskip(module)` and pytest's own version both
try to import the module to test. When this is successful, any `sys.path` or
`PATH` or `LD_LIBRARY_PATH` manipulation which that module does on import has
the potential to derail future builds causing later tests to fail (often
sporadically). Additionally, evaluating the `@importorskip` itself can fail for
psutil if ran in a background thread by pytest-xdist.

Replace the `@importorskip()` implementation with a `importlib.util.find_spec()`
based approach which *locates* the module but doesn't import it. This approach
has a couple of shortcomings:

* `find_spec()` locates submodules by first importing the top level module. That
  would defeat the purpose of this commit. As an ugly compromise, if the
  requested module is a submodule, try importing it in a subprocess to ensure
  isolation. This is much slower - where practical we should find the version
  which the submodule is introduced and use `@requires()` instead (see below).

* It can't query the package version. In some ways though this is more correct
  because packages don't have versions - distributions do. Fortunately, the
  version argument is unused both in this repo and the hooks repo so we don't
  have to worry about breaking anything by removing it.

A second similar decorator `@requires('requirement')` is added to make up for
shortcoming 2. This queries distribution metadata for versions and distribution
availability and again avoids importing anything.