Fixes for #6006 (#6026)
* Bootloader: waf: Extend the test for zlib availability.

Testing zlib-dev availability requires explicitly using libz defined
symbols instead of just compiling `include <zlib.h>` with `-lz` because
zlib's development headers may be installed but with libz.so `sstrip`-ed
making it not linkable to. This is the case on OpenWRT.

* Bootloader: waf: Allow statically linking zlib on non Windows.

The usual dynamically linking is impossible on OpenWRT which uses sstrip
on its libraries which removes headers needed by the linker. OpenWRT
can now install PyInstaller with:

    PYI_STATIC_ZLIB=1 pip install pyinstaller

Closes #6006.