Tests: Modulegraph: Remove a now unrunable test.
Python 3.10 and 3.9.6 have added a check for the recursion depths when compiling
bytecode. If the depth exceeds `sys.getrecursionlimit()`, then a recursion error
is raised. Because the test in question deliberately causes a high
recursion depth compile - we can no longer run this test.

We should have seen this sooner but it was masked by one of our tests which was
decorated with `@importorskip("IPython")`, which imports IPython, which imports
jedi, which calls `sys.setrecursionlimit(3000)`, which then allows the test to
pass again. The difference now it that our new version of `@importorskip` (see
previous commit) no longer imports the module it's testing for.

We could alternatively either conditionally xfail or skip the offending test but
such tests, given that nobody will ever look at them, invariably devolve into
clutter.