building: isolate discovery of hook directories in a subprocess (#6032)
* building: handle exceptions thrown during hook entry points loading

Limit the effect of a failing hook entry point, e.g. due to a
package with missing dependency that is not actually used by
the program.

* building: isolate discovery of hook directories in a subprocess

Perform discovery of 3rd-party hook directories in a separate
subprocess in order to avoid importing the packages that export
pyinstaller40 hook-dirs entry-points in the main process.