Improve missing linked library warnings (#6015)
* bindepend: downgrade missing library messages from errors to warnings

Downgrade missing-library messages in getImports() implementations
from errors to warnings. Missing libraries do not break build
process (and may even be harmless, depending on the situation), so
error is a bit too strong word, especially as it ends up confusing
users.

* bindepend: add warnings for libraries that cannot be found by ldd

Currently, we silently swallow "lib.so => not found" messages from
ldd, wheras we should probably be displaying a warning about those
libraries.

The behavior is likely linux-specific (or even glibc-specific),
but it's better than nothing.

* depend: generalize missing library warning suppression

Implement general missing-library warning suppression mechanism,
and apply it to codepaths for all OSes.

* depend: avoid spurious warning when libc.so points to ldd

Spotted on openwrt.

* depend: suppress warnings about missing (lib)shiboken*

These are harmless, and are result of the shiboken2/6 shared
library residing in a separate package and not being in the
search path. Therefore, PySide2/6 extensions themselves cannot
resolve the shiboken shared library, which actually ends up
loaded when the corresponding extension from shiboken package is
loaded (and this is also how PyInstaller manages to pick it up).

* depend: cleanup the import of compat.is_* members

Instead of importing each and every is_* member from compat, import
and use compat instead.