building: Complete the exclude_system_libraries feature
The last version of the commit in pr #6022 which was merged accidentally
left out the news file. This adds it back. In addition, the example in
the documentation used a format for the Analysis step that was inconsistent
with examples in the immediately following sections. This makes the format
consistent.