tests: add a new Apple Event handling test
Thoroughly test onedir and onefile bundle variant of Carbon-based
event logger, both with argv-emu enabled and disabled.

Initially, the idea was to implement different loggers to test with
(Carbon-based, tkinter-based, PyQt5-based), but that quickly devolved
into a maze of special handling for high-level toolkits' quirks.

So in the end, we test only with Carbon-based logger, which is
as low-level as it gets, intercepts raw Apple Events, and as such
offers us most insight into what is going on.