bootloader: implement support for PEP540 UTF-8 mode (#6065)
* tests: add tests for PEP540 UTF-8 mode

On python 3.7 and later, UTF-8 mode is automatically enabled if
LC_CTYPE is set to C or POSIX (macOS and linux), or if PYTHONUTF8
environment variable is set to 1 (all OSes, regardless of locale
setting). Test that this behavior is present in frozen applications
as well.

* bootloader: implement support for PEP540 UTF-8 mode

Enable PEP540 UTF-8 mode if `PYTHONUTF8` environment is set to 1
(on all OSes, regardless of locale setting). On non-Windows, enable
PEP540 UTF-8 mode for `C` and `POSIX` locales to match the behavior
of the python interpreter.

The UTF-8 mode handling code is separate from other runtime
options handling, because it seems that `Py_UTF8Mode` needs to
be set *before* `Py_SetPath()` is called to have any effect.
On the other hand, it seems that setting the warning flags via
`PySys_AddWarnOption()` in `pyi_pylib_set_runtime_opts()` needs to
be  done *after* `Py_SetPath()` is called. Therefore, UTF-8 mode
handling code is implemented as a separate utility function and
called at the beginning of `pyi_pylib_start_python()`, while
the call to `pyi_pylib_set_runtime_opts()` is left unchanged.