depend: automatically exclude Wine built-in DLLs
Automatically exclude Wine built-in DLLs, regardless of the exclude
and include lists. Wine built-in DLLs are likely to cause problems
under Windows, so excluding them increases the chances of the bundle
working under Windows. On the other hand, even with Wine DLLs excluded,
the bundle should still run under Wine as before.