Avoid collecting Wine built-in DLLs
Back-port of #6149 to the v4 branch.

* compat: implement detection of Wine environment and Wine DLLs

Add a new flag, is_win_wine, which indicates that we are running
under Wine.

Implement a new function, is_wine_dll(), which checks if the given
file is a Wine build-in DLL (PE-converted or fake/placeholder).

* depend: automatically exclude Wine built-in DLLs

Automatically exclude Wine built-in DLLs, regardless of the exclude
and include lists. Wine built-in DLLs are likely to cause problems
under Windows, so excluding them increases the chances of the bundle
working under Windows. On the other hand, even with Wine DLLs excluded,
the bundle should still run under Wine as before.

* depend: display each excluded-Wine-DLL warning only once