hooks: rework the matplotlib.backends hook (#6024)
Rework the matplotlib backend selection to support three distinct
modes, controlled by the hook config ("matplotlib" -> "backends"):
- auto-selection of used backends or the default backend
- selection of all importable backends
- selection based on user-provided list of backend names

If the config key is set to "auto", auto-detection is performed. This
is also the new default mode. It first scans the code for any uses
of `matplotlib.use()` or `mpl.use()`, and constructs the list of
corresponding backend names. If no such calls are found, the default
backend is selected, using a procedure that simulates the auto-selection
performed when `matplotlib.get_backend()` is called for the first time
 (we do not use that function directly, as it attempts to switch the
backend, which may or may not work in some corner cases, such as
our CI, or running in the headless mode). The default backend selection
amounts to selecting the first importable GUI-based backend from
pre-defined priority list: `['MacOSX', 'Qt5Agg', 'Gtk3Agg', 'TkAgg',
'WxAgg']`, and falling back to the headless `'Agg'`.

If the config key is set to "all", then old behavior that brute-force
collects all (importable) backends is enabled. The one difference is
that the import error tracebacks are not printed anymore, in order
to make the build log easier to read.

Otherwise, if the key is any other string, it is treated as a
backend name. Similarly, if it is a list, it is treated as list of
backend names.

This change aims top revent matplotlib from pulling in
every installed UI library and also give user more control over
what is collected via the config key.