hookutils: get_module_file_attribute: fix import leaks
When pkgutil.find_loader() in get_module_file_attribute() is called
with a sub-package or a sub-module name, it leaks an import of the
corresponding parent package(s).

Therefore, if the given module/package name contains a dot, we always
need to use subprocess isolation. Instead fo exec_statement, we now
use the new PyInstaller.isolated framework.

Inside the subprocess-isolated check, we need to first perform the
pkgutil-based check, as it manages to return file paths for packages
and modules that are otherwise un-importable (e.g., a C-extension
module with missing dependencies, such as PyQt5.QtMultimedia with
missing pulseaudio libs on linux).

We also import pkgutil locally, so we can eventually get rid of
module-level import, which will help us ensure that all uses of
pkgutil are properly isolated.